+++ site/pages/index.jsx
@@ -20,9 +20,11 @@ export default {
             'form': require('../../src/components/form/docs'),
             'input': require('../../src/components/input/docs'),
             'select': require('../../src/components/select/docs')
+            'radio': require('../../src/components/radio/docs')
         },
         'Outros': {
-            'card': require('../../src/components/card/docs')
+            'card': require('../../src/components/card/docs'),
+            'popover': require('../../src/components/popover/docs')
         }
     }
 }