+++ src/components/radio/docs/index.js
@@ -1,0 +1,7 @@
+import Markdown from '../../../../libs/markdown';
+
+export default class Radio extends Markdown {
+    document() {
+        return require(`./Radio.md`);
+    }
+}