+++ CHANGELOG.md
@@ -4,22 +4,24 @@ Todas as mudanças importantes deste projeto serão documentadas nesse arquivo.
 
 Formato de CHANGELOG baseado em [Keep a Changelog](https://keepachangelog.com/pt-BR/1.0.0/) seguindo o modelo de [versionamento semântico](https://semver.org/lang/pt-BR/).
 
-## [Unreleased] - 04/03/2019
+## 1.0.0 - 11/06/2019
 
 ### Adicionado
 - Componente `Aside`
 - Componente `Breadcrumb`
 - Componente `Button`
 - Componente `Card`
-- Componente `Checkbox`
 - Componente `Container`
 - Componente `Footer`
+- Componente `Form`
 - Componente `Header`
 - Componente `Input`
 - Componente `Layout`
 - Componente `Main`
-- Componente `Nav Menu`
+- Componente `Menu`
 - Componente `Step Bar`
 - Componente `Table`
 - Componente `Tag`
+- Componente `Popover`
+- Componente `Radio`
 - Componente `Select`